+++ Gopkg.lock
@@ -834,6 +834,6 @@
 [solve-meta]
   analyzer-name = "dep"
   analyzer-version = 1
-  inputs-digest = "82139b7fb2ada8f3ad396760ae1a76850e4257d6ffc2a130b02d8ab2d87796ae"
+  inputs-digest = "ebbc7f82befbe28a5a50e0d8108aad67da462ad25eb95d33ccbcfa4459a89c20"
   solver-name = "gps-cdcl"
   solver-version = 1
