+++ cloud/vsphere/constants/constants.go
@@ -1,0 +1,11 @@
+package constants
+
+const (
+	ApiServerPort                    = 443
+	VmIpAnnotationKey                = "vm-ip-address"
+	ControlPlaneVersionAnnotationKey = "control-plane-version"
+	KubeletVersionAnnotationKey      = "kubelet-version"
+	LastUpdatedKey                   = "last-updated"
+	CreateEventAction                = "Create"
+	DeleteEventAction                = "Delete"
+)
